Module Swhid

module Compute : sig ... end

A module providing various functions to compute the swhid of a given object. Supported objects are content, directory, release, revision and snapshot. The origins and visits objects are not supported.

module Download : sig ... end

Module providing functions to get a download url from a swhid.

module Json : sig ... end