swhid

swhid is an OCaml library to work with persistent identifiers found in Software Heritage, also known as swhid.

API

Swhid.Compute

Swhid.Download

Swhid.Lang

Swhid.Parser

Swhid.Pp

Private API

You shouldn't have to use any of these modules, they're used internally only.

Swhid.Git

Swhid.Json

Swhid.Lexer

Swhid.Menhir_parser