swhid

swhid is an OCaml library to work with persistent identifiers found in Software Heritage, also known as swhid.

API

Swhid.Compute A module providing various functions to compute the swhid of a given object. Supported objects are content, directory, release, revision and snapshot. The origins and visits objects are not supported.

Swhid.Download Module providing functions to get a download url from a swhid.

Swhid.Types

Swhid.Parse

Swhid.Pp

Private API

You shouldn't have to use any of these modules, they're used internally only.

Swhid.Json

Swhid.Lexer

Swhid.Menhir_parser